fix: reworked containerdriver struct
